# minishop-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MiniShop API
  version: 2.1.0
  description: Published contract last reviewed 2025-11-18. It intentionally trails selected source behavior for audit-workflow testing.
servers:
  - url: https://api.staging.minishop.invalid/api
security:
  - bearerAuth: []
paths:
...
  /v1/orders:
    get:
      summary: List caller orders
      responses:
        '200': {description: Orders}
    post:
      summary: Submit order
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OrderWrite'}}}}
      responses:
        '201': {description: Created}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/orders/{orderId}:
    get:
      summary: Read order
      parameters: [{$ref: '#/components/parameters/OrderId'}]
      responses:
        '200': {description: Order, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/orders/{orderId}/status:
    patch:
      summary: Change fulfillment state
      parameters: [{$ref: '#/components/parameters/OrderId'}]
      requestBody: {content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: string, enum: [pending,paid,shipped,delivered,cancelled]}}}}}}
      responses:
        '200': {description: Updated}
...
  /v1/cart/items:
    post:
      summary: Add cart item
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CartItemWrite'}}}}
      responses:
        '201': {description: Added}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /v1/inventory/{productId}:
    get:
      summary: Product inventory
      parameters: [{$ref: '#/components/parameters/ProductId'}]
      responses:
        '200': {description: Inventory}
  /v1/inventory/{productId}/adjustments:
    post:
      summary: Adjust inventory
      parameters: [{$ref: '#/components/parameters/ProductId'}]
      requestBody: {content: {application/json: {schema: {type: object, required: [delta], properties: {delta: {type: integer}, reasonDetail: {type: string}}}}}}
      responses:
        '201': {description: Adjusted}
  /v1/products/{productId}/reviews:
    get:
      summary: Approved reviews
      parameters: [{$ref: '#/components/parameters/ProductId'}]
      responses:
        '200': {description: Reviews}
    post:
      summary: Submit review
      parameters: [{$ref: '#/components/parameters/ProductId'}]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/ReviewWrite'}}}}
      responses:
        '201': {description: Submitted}
  /v1/product/{id}:
    get:
      deprecated: true
      security: []
      summary: Legacy product lookup
      parameters: [{in: path, name: id, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Product, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    ProductId: {in: path, name: productId, required: true, schema: {type: string, format: uuid}}
    OrderId: {in: path, name: orderId, required: true, schema: {type: string, format: uuid}}
...
  schemas:
...
    ProductWrite:
      type: object
      required: [sku,name,pricePaise,categoryId]
      properties:
        sku: {type: string}
        name: {type: string, minLength: 3, maxLength: 120}
        description: {type: string}
        pricePaise: {type: integer, minimum: 1}
        categoryId: {type: string, format: uuid}
        tags: {type: array, items: {type: string}}
    Product:
      allOf:
        - {$ref: '#/components/schemas/ProductWrite'}
        - {type: object, required: [id], properties: {id: {type: string, format: uuid}, active: {type: boolean}}}
...
    CartItemWrite:
      type: object
      required: [productId,quantity]
      properties: {productId: {type: string, format: uuid}, quantity: {type: integer, minimum: 1, maximum: 99}}
    OrderWrite:
      type: object
      required: [items]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/CartItemWrite'}}, addressId: {type: string, format: uuid}}
    Order:
      type: object
      properties: {id: {type: string, format: uuid}, orderNumber: {type: string}, status: {type: string}, totalPaise: {type: integer}}
...
    ReviewWrite:
      type: object
      required: [rating,comment]
      properties: {rating: {type: integer, minimum: 1, maximum: 5}, comment: {type: string}}
    Error:
      type: object
      properties: {code: {type: string}, message: {type: string}, details: {type: array, items: {type: object}}}
  responses:
...
    NotFound: {description: Resource absent, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Conflict: {description: State conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}